AudioIO does not depend on ProjectSettings...
... Rewrite event handling for change of play speed.

This resolves conflict of recent developments with future extraction of
lib-audioio
